Packer hypervbox (#61)
* GH-55

+++ Source/Invoke-PackerBuild.ps1
@@ -1,9 +1,10 @@
+[CmdletBinding(SupportsShouldProcess=$true)]
 param(
     [Parameter(Mandatory=$true,ParameterSetName="AWS EC2 AMI")]
     [Parameter(Mandatory=$true,ParameterSetName="Vagrant Hyper-V")]
     [string]$ISHVersion,
     [Parameter(Mandatory=$false,ParameterSetName="AWS EC2 AMI")]
-    [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
+    [Parameter(Mandatory=$true,ParameterSetName="Vagrant Hyper-V")]
     [string]$MockConnectionString=$null,
     [Parameter(Mandatory=$false,ParameterSetName="AWS EC2 AMI")]
     [string]$SourceAMI,
@@ -26,7 +27,16 @@ param(
     [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
     [string]$SwitchName="External Virtual Switch",
     [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
-    [string]$BoxPath="$($env:TEMP)\ISH.$ISHVersion-hyperv-iso.box"
+    [ValidateSet('2012_R2', '2016')]
+    [string]$ServerVersion="2016",
+    [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
+    [string]$OutputPath="$env:TEMP",
+    [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
+    [switch]$NoWindowsUpdates=$false,
+    [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
+    [switch]$ServerCore=$false,
+    [Parameter(Mandatory=$false,ParameterSetName="Vagrant Hyper-V")]
+    [switch]$Force=$false
 )
 
 if ($PSBoundParameters['Debug']) {
@@ -109,6 +119,45 @@ switch ($PSCmdlet.ParameterSetName) {
         $logRegExSource="amazon-ebs"
     }
     'Vagrant Hyper-V' {
+
+        $boxNameSegments=@(
+            "windowsserver"
+            $ServerVersion
+            "ish.$ISHVersion"
+        )
+        $autounattendFolderSegments=@(
+            $ServerVersion
+        )
+        if ($ServerCore)
+        {
+            $autounattendFolderSegments+="core"
+            $boxNameSegments+="core"
+        }
+		
+        if ($NoWindowsUpdates)
+        {
+            $autounattendFolderSegments+="no_windows_updates"
+            $boxNameSegments+="no_windows_updates"
+        }
+	
+        $autounattendFolder=$autounattendFolderSegments -join '_'
+
+        $boxNameSegments+="hyperv"
+        $boxName=$boxNameSegments -join '_'
+        $boxPath=Join-Path $OutputPath "$($boxName).box"
+        if(Test-Path -Path $boxPath)
+        {
+            if($Force)
+            {
+                Write-Warning "Removing $boxPath"
+            }
+            else
+            {
+                Write-Error "Box $boxPath already exists"
+                return 1
+            }
+        }
+
         $packerArgs+=@(
             "-var"
             "iso_url=$ISOUrl"
@@ -123,13 +172,25 @@ switch ($PSCmdlet.ParameterSetName) {
             "-var"
             "aws_secret_key=$SecretKey"
             "-var"
-            "output_box_path=$BoxPath"
+            "output_box_path=$boxPath"
+            "-var"
+            "autounattend_folder=$autounattendFolder"
         )
 
+        if($MockConnectionString)
+        {
+            $packerArgs+=@(
+                "-var"
+                "ish_mock_connectionstring=$MockConnectionString"
+	        )
+        }
 
-        $packerFileNameName="ish-HyperV-Vagrant.json"
+        $packerFileNameName="ish-$ServerVersion-vagrant-hyperv-iso.json"
+        $logRegExSource="hyperv-iso"
     }
 }
+
+Write-Host "Using $packerFileNameName"
 
 $packerArgs+=$packerFileNameName
 
@@ -137,36 +198,48 @@ Push-Location -Path "$PSScriptRoot\Packer" -StackName Packer
 
 try
 {
-    $env:PACKER_LOG=1
-    $packetLogPath=Join-Path $env:TEMP "$($packerFileNameName).txt"
-    if(Test-Path -Path $packetLogPath)
-    {
-        Remove-Item -Path $packetLogPath -Force
+    $invokedPacker=$false
+
+    if ($PSCmdlet.ShouldProcess($packerFileNameName, "packer build")){
+        $invokedPacker=$true
+        $env:PACKER_LOG=1
+        $packetLogPath=Join-Path $env:TEMP "$($packerFileNameName).txt"
+        if(Test-Path -Path $packetLogPath)
+        {
+            Remove-Item -Path $packetLogPath -Force
+        }
+        $env:PACKER_LOG_PATH=$packetLogPath
+        Write-Host "packer $packerArgs"
+        & packer $packerArgs
+        Write-Host "LASTEXITCODE=$LASTEXITCODE"
     }
-    $env:PACKER_LOG_PATH=$packetLogPath
-    Write-Host "packer $packerArgs"
-    & packer $packerArgs
-    Write-Host "LASTEXITCODE=$LASTEXITCODE"
+    else
+    {
+        Write-Host "packer $($packerArgs -join ' ')"
+    }
 }
 finally
 {
-    Write-Warning "Packer log file available in $packetLogPath"
-    Pop-Location -StackName Packer
-
-    if($LASTEXITCODE -ne 0)
+    if($invokedPacker)
     {
-        if($logRegExSource)
+        Write-Warning "Packer log file available in $packetLogPath"
+        Pop-Location -StackName Packer
+
+        if($LASTEXITCODE -ne 0)
         {
-            $packerLogContent=Get-Content -Path  $packetLogPath -Raw
-            $regex=".*$($logRegExSource): (?<Objs>\<Objs.*\</Objs\>).*"
-            $matchCollections=[regex]::Matches($packerLogContent,$regex)
-            if($matchCollections.Count -gt 0)
+            if($logRegExSource)
             {
-                Write-Warning "Packer Objs xml entries available:"
-                for($i=0;$i -lt $matchCollections.Count;$i++) {
-                    $objsItemPath=$packetLogPath.Replace(".txt",".$i.xml")
-                    $matchCollections[$i].Groups['Objs'].Value | Format-TidyXml | Out-File -FilePath $objsItemPath
-                    Write-Warning $objsItemPath
+                $packerLogContent=Get-Content -Path  $packetLogPath -Raw
+                $regex=".*$($logRegExSource): (?<Objs>\<Objs.*\</Objs\>).*"
+                $matchCollections=[regex]::Matches($packerLogContent,$regex)
+                if($matchCollections.Count -gt 0)
+                {
+                    Write-Warning "Packer Objs xml entries available:"
+                    for($i=0;$i -lt $matchCollections.Count;$i++) {
+                        $objsItemPath=$packetLogPath.Replace(".txt",".$i.xml")
+                        $matchCollections[$i].Groups['Objs'].Value | Format-TidyXml | Out-File -FilePath $objsItemPath
+                        Write-Warning $objsItemPath
+                    }
                 }
             }
         }
